Type resilience: Guard against int v. double arithmetic
This commit makes the makeLayeredGrid resilient against various
integer arithmetic type errors.  By explicitly using DOUBLE, we
automatically handle types that come from (e.g.,) MEX-ed grid
constructors (like 'mex_triangleGrid').

+++ gridprocessing/makeLayeredGrid.m
@@ -53,7 +53,7 @@ G.type = [G.type, { mfilename }];
 
 % Faces with horizontal normal
 %-----------------------------
-edges = reshape(G.faces.nodes(:,1), 2, []) .';
+edges = reshape(double(G.faces.nodes(:,1)), 2, []) .';
 dz    = G.nodes.num;
 p1    = repmat(edges(:,1), [nlayers,1])+...
         kron((0:nlayers-1)', ones(size(edges,1), 1))*dz;
@@ -76,7 +76,7 @@ hNeighbors(bdry)=0;
 % Faces with vertical normal
 %----------------------------
 cn     = getCellNodes(G);
-vFaces = repmat(cn, [nlayers+1, 1]) + ...
+vFaces = repmat(double(cn), [nlayers + 1, 1]) + ...
          kron((0:nlayers)', ones(size(cn,1), 1))*dz;
 vNumNodes = repmat(diff(G.cells.facePos), [nlayers+1, 1]);
